renamed first-pass-canonical-image to pick-canonical-image
