+++ client/lib/post-normalizer/rule-pick-canonical-image.js
