+++ client/state/reader/posts/normalization-rules.js
@@ -23,7 +23,7 @@ import removeStyles from 'lib/post-normalizer/rule-content-remove-styles';
 import makeImagesSafe from 'lib/post-normalizer/rule-content-make-images-safe';
 import { disableAutoPlayOnMedia, disableAutoPlayOnEmbeds } from 'lib/post-normalizer/rule-content-disable-autoplay';
 import decodeEntities from 'lib/post-normalizer/rule-decode-entities';
-import pickCanonicalImage from 'lib/post-normalizer/rule-first-pass-canonical-image';
+import pickCanonicalImage from 'lib/post-normalizer/rule-pick-canonical-image';
 import makeSiteIdSafeForApi from 'lib/post-normalizer/rule-make-site-id-safe-for-api';
 import pickPrimaryTag from 'lib/post-normalizer/rule-pick-primary-tag';
 import preventWidows from 'lib/post-normalizer/rule-prevent-widows';
@@ -170,7 +170,6 @@ const fastPostNormalizationRules = flow( [
 		detectMedia,
 		detectPolls,
 	] ),
-	firstPassCanonicalImage,
 	config.isEnabled( 'reader/refresh/stream' ) ? createBetterExcerptRefresh : createBetterExcerpt,
 	pickCanonicalMedia,
 	classifyPost,
